cinema_sauro/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
psycopg2-binary==2.9.5